requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1

# ── GitHub integration ────────────────────────────────────────────────────────
PyGithub==2.5.0

# ── AI providers ──────────────────────────────────────────────────────────────
# Anthropic SDK — used when AI_PROVIDER=anthropic (default)
anthropic==0.40.0
# OpenAI SDK — used when AI_PROVIDER=qwen (DashScope is OpenAI-compatible)
openai==1.57.0

# ── Async task queue ──────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# ── HTTP client (used in dashboard state) ─────────────────────────────────────
httpx==0.27.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-anyio==0.0.0
anyio[trio]==4.6.2
pytest-asyncio==0.24.0
